+++ src/vuex/account.module.js
@@ -71,16 +71,16 @@ export const getters = {
   [vuexTypes.account]: state => state.account,
   [vuexTypes.accountId]: state => state.account.id,
   [vuexTypes.accountBalances]: state => state.balancesDetails
-    .map(item => new BalanceRecord(item)),
+    .map(item => new BalanceRecord(item, item.asset.trailingDigits)),
   [vuexTypes.accountBalancesByOwner]: (state, getters) => owner => {
     return getters[vuexTypes.accountBalances]
       .filter(item => item.asset.owner === owner)
   },
   [vuexTypes.accountOwnedAssetsBalances]: state => state.balancesDetails
-    .map(item => new BalanceRecord(item))
+    .map(item => new BalanceRecord(item, item.asset.trailingDigits))
     .filter(i => i.asset.owner === state.account.id) || {},
   [vuexTypes.accountBalanceByCode]: state => code => state.balancesDetails
-    .map(item => new BalanceRecord(item))
+    .map(item => new BalanceRecord(item, item.asset.trailingDigits))
     .find(i => i.asset.code === code) || {},
   [vuexTypes.accountRoleId]: state => Number(
     _get(state.account, 'role.id')